Create set up for flow bug on iOS
In this change, we create a class CoroutineLauncher that launches two coroutines with two different scopes. Each coroutine collects on a flow provided by a swift class that inherits from a shared kotlin class. The code exposes a button to allow the user to cancel one of the two scopes. However, even after canceling one of the scopes, if we emit from the provided flow, both coroutines will activate. As a side note, if we provide a flow that does not explicitely inherit from UnitFlow, this bug doesn't occur. See the ConcreteUnitFlowNoInterface class. As a second side note, this repo is using the native-mt version of the coroutines library, but the bug still occurs if we did not use the native-mt version and implemented our own iOS dispatcher.